- Make use of new NationPreTownLeaveEvent
- Create second listener for non-Action events.
