+++ src/com/palmergames/bukkit/towny/war/siegewar/listeners/SiegeWarActionListener.java
@@ -2,41 +2,24 @@ package com.palmergames.bukkit.towny.war.siegewar.listeners;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
-
-import org.bukkit.Material;
 import org.bukkit.block.Block;
-import org.bukkit.block.BlockFace;
 import org.bukkit.event.EventHandler;
-import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
-import org.bukkit.event.block.BlockPistonExtendEvent;
-import org.bukkit.event.block.BlockPistonRetractEvent;
-import org.bukkit.event.player.PlayerItemConsumeEvent;
-
 import com.palmergames.bukkit.towny.Towny;
-import com.palmergames.bukkit.towny.TownyEconomyHandler;
-import com.palmergames.bukkit.towny.TownyMessaging;
 import com.palmergames.bukkit.towny.TownySettings;
-import com.palmergames.bukkit.towny.TownyUniverse;
-import com.palmergames.bukkit.towny.event.NationPreAddTownEvent;
-import com.palmergames.bukkit.towny.event.PreDeleteNationEvent;
 import com.palmergames.bukkit.towny.event.actions.TownyBuildEvent;
 import com.palmergames.bukkit.towny.event.actions.TownyBurnEvent;
 import com.palmergames.bukkit.towny.event.actions.TownyDestroyEvent;
 import com.palmergames.bukkit.towny.event.actions.TownyExplodingBlocksEvent;
-import com.palmergames.bukkit.towny.event.nation.PreNewNationEvent;
 import com.palmergames.bukkit.towny.event.player.PlayerKilledPlayerEvent;
-import com.palmergames.bukkit.towny.object.Nation;
 import com.palmergames.bukkit.towny.object.Translation;
-import com.palmergames.bukkit.towny.utils.TownPeacefulnessUtil;
 import com.palmergames.bukkit.towny.war.siegewar.SiegeWarDeathController;
-import com.palmergames.bukkit.towny.war.siegewar.objects.Siege;
 import com.palmergames.bukkit.towny.war.siegewar.utils.SiegeWarBlockUtil;
 import com.palmergames.bukkit.towny.war.siegewar.utils.SiegeWarDistanceUtil;
 
 public class SiegeWarActionListener implements Listener {
 
+	@SuppressWarnings("unused")
 	private final Towny plugin;
 	
 	public SiegeWarActionListener(Towny instance) {
@@ -83,35 +66,6 @@ public class SiegeWarActionListener implements Listener {
 	}
 	
 	/*
-	 * SW will prevent someone in a banner area from curing their poisoning with milk.
-	 */
-	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
-	public void onPlayerConsume(PlayerItemConsumeEvent event) {
-		if (plugin.isError()) {
-			event.setCancelled(true);
-			return;
-		}
-
-		if(TownySettings.getWarSiegeEnabled()) {
-			try {
-				//Prevent milk bucket usage while attempting to gain banner control
-				if(event.getItem().getType() == Material.MILK_BUCKET) {
-					for(Siege siege: TownyUniverse.getInstance().getDataSource().getSieges()) {
-						if(siege.getBannerControlSessions().containsKey(event.getPlayer())) {
-							event.setCancelled(true);
-							TownyMessaging.sendErrorMsg(event.getPlayer(), Translation.of("msg_war_siege_zone_milk_bucket_forbidden_while_attempting_banner_control"));
-						}
-					}
-				}
-		
-			} catch (Exception e) {
-				System.out.println("Problem evaluating siege player consume event");
-				e.printStackTrace();
-			}
-		}
-	}
-	
-	/*
 	 * SW can affect the emptying of buckets, which could affect a banner.
 	 */
 	@EventHandler
@@ -139,119 +93,5 @@ public class SiegeWarActionListener implements Listener {
 			SiegeWarDeathController.evaluateSiegePlayerDeath(event.getVictim(), event.getPlayerDeathEvent());
 		}
 	}
-	
-	/*
-	 * SW limits which Towns can join or be added to a nation.
-	 */
-	@EventHandler
-	public void onNationAddTownEvent(NationPreAddTownEvent event) {
-		if(TownySettings.getWarSiegeEnabled() && TownySettings.getWarCommonPeacefulTownsEnabled() && event.getTown().isPeaceful()) {
-			Set<Nation> validGuardianNations = TownPeacefulnessUtil.getValidGuardianNations(event.getTown());
-			if(!validGuardianNations.contains(event.getNation())) {
-				event.setCancelMessage(Translation.of("msg_war_siege_peaceful_town_cannot_join_nation", 
-						event.getTown().getName(),
-						event.getNation().getName(),
-						TownySettings.getWarSiegePeacefulTownsGuardianTownMinDistanceRequirement(),
-						TownySettings.getWarSiegePeacefulTownsGuardianTownPlotsRequirement()));
-				event.setCancelled(true);
-			}
-		}
-	}
-	
-	/*
-	 * SW warns peaceful towns who make nations their decision may be a poor one, but does not stop them.
-	 */
-	@EventHandler
-	public void onNewNationEvent(PreNewNationEvent event) {
-		if (TownySettings.getWarSiegeEnabled() && TownySettings.getWarCommonPeacefulTownsEnabled()
-				&& event.getTown().isPeaceful()) {
-			TownyMessaging.sendMsg(event.getTown().getMayor(),
-					Translation.of("msg_war_siege_warning_peaceful_town_should_not_create_nation"));
-		}
-	}
-	
-	/*
-	 * SW will warn a nation about to delete itself that it can claim a refund after the fact.
-	 */
-	@EventHandler
-	public void onNationDeleteEvent(PreDeleteNationEvent event) {
-		//If nation refund is enabled, warn the player that they will get a refund (and indicate how to claim it).
-		if (TownySettings.getWarSiegeEnabled() && TownySettings.isUsingEconomy()
-				&& TownySettings.getWarSiegeRefundInitialNationCostOnDelete()) {
-			int amountToRefund = (int)(TownySettings.getNewNationPrice() * 0.01 * TownySettings.getWarSiegeNationCostRefundPercentageOnDelete());
-			TownyMessaging.sendMsg(event.getNation().getKing(), Translation.of("msg_err_siege_war_delete_nation_warning", TownyEconomyHandler.getFormattedBalance(amountToRefund)));
-		}
 
-	}
-	
-	/*
-	 * Duplicates what exists in the TownyBlockListener but on a higher priority.
-	 */
-	@EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
-	public void onBlockPistonRetract(BlockPistonRetractEvent event) {
-
-		if (plugin.isError()) {
-			event.setCancelled(true);
-			return;
-		}
-
-		if (testBlockMove(event.getBlock(), event.isSticky() ? event.getDirection().getOppositeFace() : event.getDirection()))
-			event.setCancelled(true);
-
-		List<Block> blocks = event.getBlocks();
-		
-		if (!blocks.isEmpty()) {
-			//check each block to see if it's going to pass a plot boundary
-			for (Block block : blocks) {
-				if (testBlockMove(block, event.getDirection()))
-					event.setCancelled(true);
-			}
-		}
-	}
-
-	/*
-	 * Duplicates what exists in the TownyBlockListener but on a higher priority.
-	 */
-	@EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
-	public void onBlockPistonExtend(BlockPistonExtendEvent event) {
-
-		if (plugin.isError()) {
-			event.setCancelled(true);
-			return;
-		}
-		
-		if (testBlockMove(event.getBlock(), event.getDirection()))
-			event.setCancelled(true);
-		
-		List<Block> blocks = event.getBlocks();
-
-		if (!blocks.isEmpty()) {
-			//check each block to see if it's going to pass a plot boundary
-			for (Block block : blocks) {
-				if (testBlockMove(block, event.getDirection()))
-					event.setCancelled(true);
-			}
-		}
-	}
-
-	/**
-	 * Decides whether blocks moved by pistons follow the rules.
-	 * 
-	 * @param block - block that is being moved.
-	 * @param direction - direction the piston is facing.
-	 * 
-	 * @return true if block is able to be moved according to siege war rules. 
-	 */
-	private boolean testBlockMove(Block block, BlockFace direction) {
-
-		Block blockTo = block.getRelative(direction);
-
-		if(TownySettings.getWarSiegeEnabled()) {
-			if(SiegeWarBlockUtil.isBlockNearAnActiveSiegeBanner(block) || SiegeWarBlockUtil.isBlockNearAnActiveSiegeBanner(blockTo)) {
-				return true;
-			}
-		}
-
-		return false;
-	}
 }
